rename webview to mattermost-observer, and remove `checkActivity` for now
